feat: fetch reference data for dataset form
