+++ apps/dataset-catalog/app/catalogs/[catalogId]/datasets/new/page.tsx
@@ -3,13 +3,23 @@ import { DatasetForm } from '../../../../../components/dataset-form';
 import { datasetToBeCreatedTemplate } from '../../../../../components/dataset-form/dataset-initial-values';
 import { Params } from 'next/dist/shared/lib/router/utils/route-matcher';
 import { getTranslateText, localization } from '@catalog-frontend/utils';
-import { Organization } from '@catalog-frontend/types';
-import { getOrganization } from '@catalog-frontend/data-access';
+import { DataTheme, LosTheme, Organization, ReferenceDataCode } from '@catalog-frontend/types';
+import { getDatasetTypes, getDataThemes, getLosThemes, getOrganization } from '@catalog-frontend/data-access';
 
 export default async function NewDatasetPage({ params }: Params) {
   const initialValues = datasetToBeCreatedTemplate();
   const { catalogId } = params;
   const organization: Organization = await getOrganization(catalogId).then((res) => res.json());
+
+  const [losThemesResponse, dataThemesResponse, datasetTypesResponse] = await Promise.all([
+    getLosThemes().then((res) => res.json()),
+    getDataThemes().then((res) => res.json()),
+    getDatasetTypes().then((res) => res.json()),
+  ]);
+
+  const losThemes: LosTheme[] = losThemesResponse.losNodes;
+  const dataThemes: DataTheme[] = dataThemesResponse.dataThemes;
+  const datasetTypes: ReferenceDataCode[] = datasetTypesResponse.datasetTypes;
 
   const breadcrumbList = [
     {
@@ -33,6 +43,9 @@ export default async function NewDatasetPage({ params }: Params) {
         <DatasetForm
           initialValues={initialValues}
           submitType={'create'}
+          losThemes={losThemes}
+          dataThemes={dataThemes}
+          datasetTypes={datasetTypes}
         ></DatasetForm>
       </div>
     </>
