+++ apps/dataset-catalog/components/dataset-form/dataset-form.module.css
@@ -3,9 +3,3 @@
   flex-direction: column;
   gap: 0.5rem;
 }
-
-.spinner {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
